improvise error handling

+++ backend/src/services/pod.service.ts
@@ -1,3 +1,4 @@
+import { AppError } from "../utils/AppError";
 import logger from "../utils/logger";
 import { K8sService } from "./k8s.service";
 import { NamespaceService } from "./namespace.service";
@@ -14,7 +15,11 @@ export class PodService extends K8sService {
 
     if (namespace) {
       
-      await this.namespaceService.checkNamespaceExists(namespace);
+      const namespaceExists = await this.namespaceService.checkNamespaceExists(namespace);
+
+      if (!namespaceExists) {
+        throw new AppError(`Namespace ${namespace} does not exist`, 404);
+      }
 
       const res = await this.coreApi.listNamespacedPod({ namespace });
       logger.info(`Fetched ${res.items.length} pods from namespace ${namespace}`);
@@ -26,9 +31,31 @@ export class PodService extends K8sService {
     return res.items;
   }
 
+  async checkPodExists(podName: string, namespace: string) {
+
+    const namespaceExists = await this.namespaceService.checkNamespaceExists(namespace);
+
+    if (!namespaceExists) {
+      throw new AppError(`Namespace ${namespace} does not exist`, 404);
+    }
+
+    const pods = await this.getPods(namespace);
+    const podExists = pods.filter(p => p.metadata?.name === podName).length > 0;
+
+    return !podExists ? false : true;
+  }
+
   async getPodLogs({ podName, namespace, container }: { podName: string; namespace: string; container?: string }) {
 
-    await this.namespaceService.checkNamespaceExists(namespace);
+    const namespaceExists = await this.namespaceService.checkNamespaceExists(namespace);
+    if (!namespaceExists) {
+      throw new AppError(`Namespace ${namespace} does not exist`, 404);
+    }
+
+    const podExists = await this.checkPodExists(podName, namespace);
+    if (!podExists) {
+      throw new AppError(`Pod ${podName} does not exist in namespace ${namespace}`, 404);
+    }
 
     const res = await this.coreApi.readNamespacedPodLog({
       name: podName,
@@ -40,8 +67,16 @@ export class PodService extends K8sService {
   }
 
   async getPodDetails(podName: string, namespace: string) {
+    
+    const namespaceExists = await this.namespaceService.checkNamespaceExists(namespace);
+    if (!namespaceExists) {
+      throw new AppError(`Namespace ${namespace} does not exist`, 404);
+    } 
 
-    await this.namespaceService.checkNamespaceExists(namespace);
+    const podExists = await this.checkPodExists(podName, namespace);
+    if (!podExists) {
+      throw new AppError(`Pod ${podName} does not exist in namespace ${namespace}`, 404);
+    }
 
     const res = await this.coreApi.readNamespacedPod({
       name: podName,
@@ -52,8 +87,16 @@ export class PodService extends K8sService {
 
   async deletePod(podName: string, namespace: string) {
     
-    await this.namespaceService.checkNamespaceExists(namespace);
-    
+    const namespaceExists = await this.namespaceService.checkNamespaceExists(namespace);
+    if (!namespaceExists) {
+      throw new AppError(`Namespace ${namespace} does not exist`, 404);
+    }
+
+    const podExists = await this.checkPodExists(podName, namespace);
+    if (!podExists) {
+      throw new AppError(`Pod ${podName} does not exist in namespace ${namespace}`, 404);
+    }
+
     const res = await this.coreApi.deleteNamespacedPod({
       name: podName,
       namespace
